continue tidying/splitting up workspaceapps code
